Move around to avoid circular dependency
