+++ constructor/executor/types.go
@@ -12,32 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package constructor
+package executor
 
 import (
 	"context"
 	"time"
 
 	"github.com/coinbase/rosetta-sdk-go/keys"
-	"github.com/coinbase/rosetta-sdk-go/parser"
-	"github.com/coinbase/rosetta-sdk-go/storage"
 	"github.com/coinbase/rosetta-sdk-go/types"
 )
 
 const (
-	// NoHeadBlockWaitTime is the amount of
-	// time we wait when no blocks have been
-	// synced.
-	NoHeadBlockWaitTime = 1 * time.Second
-
 	// BalanceWaitTime is the amount of time
 	// we wait between balance checks.
 	BalanceWaitTime = 5 * time.Second
-
-	// NoJobsWaitTime is the amount of time
-	// we wait when no jobs are available
-	// to process.
-	NoJobsWaitTime = 10 * time.Second
 )
 
 // ReservedVariable is a reserved variable
@@ -323,45 +311,12 @@ type Helper interface {
 		*keys.KeyPair,
 	) error
 
-	// AllAddresses returns a slice of all known addresses.
-	AllAddresses(ctx context.Context) ([]string, error)
-
-	// LockedAccounts is a slice of all addresses currently sending or receiving
-	// funds.
-	LockedAddresses(context.Context) ([]string, error)
-
 	// Balance returns the balance
 	// for a provided address.
 	Balance(context.Context, *types.AccountIdentifier) ([]*types.Amount, error)
 
 	// Coins returns all *types.Coin owned by an address.
 	Coins(context.Context, *types.AccountIdentifier) ([]*types.Coin, error)
-
-	// BroadcastAll broadcasts all transactions considered ready for
-	// broadcast (unbroadcasted or stale).
-	BroadcastAll(context.Context) error
-
-	// HeadBlockExists returns a boolean indicating if a block
-	// has been synced by BlockStorage.
-	HeadBlockExists(context.Context) bool
-
-	// DatabaseTransaction returns a new storage.DatabaseTransaction.
-	// This is used to update jobs and enque them for broadcast atomically.
-	DatabaseTransaction(context.Context) storage.DatabaseTransaction
-
-	// AllBroadcasts returns a slice of all in-progress broadcasts.
-	AllBroadcasts(ctx context.Context) ([]*storage.Broadcast, error)
-
-	// Broadcast enqueues a particular intent for broadcast.
-	Broadcast(
-		context.Context,
-		storage.DatabaseTransaction,
-		string, // Job.Identifier
-		*types.NetworkIdentifier,
-		[]*types.Operation,
-		*types.TransactionIdentifier,
-		string, // network transaction
-	) error
 
 	// Derive returns a new address for a provided publicKey.
 	Derive(
@@ -371,103 +326,15 @@ type Helper interface {
 		map[string]interface{},
 	) (string, map[string]interface{}, error)
 
-	// Preprocess calls the /construction/preprocess endpoint
-	// on an offline node.
-	Preprocess(
-		context.Context,
-		*types.NetworkIdentifier,
-		[]*types.Operation,
-		map[string]interface{},
-	) (map[string]interface{}, error)
+	// AllAddresses returns a slice of all known addresses.
+	AllAddresses(ctx context.Context) ([]string, error)
 
-	// Metadata calls the /construction/metadata endpoint
-	// using the online node.
-	Metadata(
-		context.Context,
-		*types.NetworkIdentifier,
-		map[string]interface{},
-	) (map[string]interface{}, error)
-
-	// Payloads calls the /construction/payloads endpoint
-	// using the offline node.
-	Payloads(
-		context.Context,
-		*types.NetworkIdentifier,
-		[]*types.Operation,
-		map[string]interface{},
-	) (string, []*types.SigningPayload, error)
-
-	// Parse calls the /construction/parse endpoint
-	// using the offline node.
-	Parse(
-		context.Context,
-		*types.NetworkIdentifier,
-		bool, // signed
-		string, // transaction
-	) ([]*types.Operation, []string, map[string]interface{}, error)
-
-	// Combine calls the /construction/combine endpoint
-	// using the offline node.
-	Combine(
-		context.Context,
-		*types.NetworkIdentifier,
-		string, // unsigned transaction
-		[]*types.Signature,
-	) (string, error)
-
-	// Hash calls the /construction/hash endpoint
-	// using the offline node.
-	Hash(
-		context.Context,
-		*types.NetworkIdentifier,
-		string, // network transaction
-	) (*types.TransactionIdentifier, error)
-
-	// Sign returns signatures for the provided
-	// payloads.
-	Sign(
-		context.Context,
-		[]*types.SigningPayload,
-	) ([]*types.Signature, error)
-}
-
-// JobStorage allows for the persistent and transactional
-// storage of Jobs.
-type JobStorage interface {
-	// Ready returns the jobs that are ready to be processed.
-	Ready(context.Context) ([]*Job, error)
-
-	// Processing returns the number of jobs processing
-	// for a particular workflow.
-	Processing(context.Context, string) (int, error)
-
-	// Update stores an updated *Job in storage
-	// and returns its UUID (which won't exist
-	// on first update).
-	Update(context.Context, storage.DatabaseTransaction, *Job) (string, error)
-
-	// Get fetches a *Job by Identifier. It returns an error
-	// if the identifier doesn't exist.
-	Get(context.Context, storage.DatabaseTransaction, string) (*Job, error)
+	// LockedAccounts is a slice of all addresses currently sending or receiving
+	// funds.
+	LockedAddresses(context.Context) ([]string, error)
 }
 
 // Worker processes jobs.
 type Worker struct {
 	helper Helper
-}
-
-// Coordinator faciliates the creation and processing
-// of jobs.
-type Coordinator struct {
-	storage JobStorage
-	helper  Helper
-	parser  *parser.Parser
-
-	attemptedJobs        []string
-	attemptedWorkflows   []string
-	seenErrCreateAccount bool
-
-	workflows             []*Workflow
-	createAccountWorkflow *Workflow
-	requestFundsWorkflow  *Workflow
 }
